Check the existence of old files before reverting

+++ patches/revert_helper.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+if [ ! -f /boot/$BZFILE.old ] || [ ! -f /boot/$BZFILE.sha256.old ]; then
+    echo "No old files found. Nothing to revert."
+    exit 0
+fi
+
 mv -f /boot/$BZFILE.old /boot/$BZFILE
 mv -f /boot/$BZFILE.sha256.old /boot/$BZFILE.sha256
